completed user and creating channel

+++ Backend/config/cloudinary.js
@@ -3,8 +3,8 @@ import fs from 'fs'
 
 const uploadOnCloudinary = async(filePath) => {
     cloudinary.config({ 
-        cloud_name: process.env.CLOUDINARY_NAME, 
-        api_key: process.env.CLOUDINARY_API_KEY, 
+        cloud_name: process.env.CLOUDINARY_NAME,
+        api_key: process.env.CLOUDINARY_API_KEY,
         api_secret: process.env.CLOUDINARY_SECRET_KEY
     });
 
@@ -13,10 +13,10 @@ const uploadOnCloudinary = async(filePath) => {
         const result = await cloudinary.uploader.upload(filePath, {resource_type : 'auto'});
         return result.secure_url;
     } catch (error) {
-        console.log(`some error occured while uploading to cloudinay : ${error}`);
+        console.log('some error occured while uploading to cloudinay :', error);
     }
     finally{
-        fs.unlinkSync(filePath);
+        if(filePath && fs.existsSync(filePath))    fs.unlinkSync(filePath);
     }
 }
